shaping.lua: use new tsa endpoint, falling back to old
Adjust our subscription to try the new generic event endpoint,
and if that 404's, fall back to the suspension endpoint.

This commit doesn't add support for the new bounce records,
it is focused on ensuring that suspension information is
returned correctly for both old and new endpoints.

The integration tests will use the new endpoint only
because it is too much of a faff to run them against
a legacy tsa-daemon (since we can only build and run
the current one here).

I tested this commit in phases though:

* Deliberately used the wrong URL for the new endpoint to
  trigger the 404 logic case and handle fallback
* Inserted an error in the fallback case to ensure that
  the fallback wasn't happening once the new endpoint URL
  was set to the correct one.

So this should enable folks to upgrade either TSA or kumod
in either order across their deployment, without the "mismatch"
in TSA and kumod versions harming their traffic.

refs: https://github.com/KumoCorp/kumomta/issues/272