log-types: reduce size of JsonLogRecord by ~30% in common case
One of the things that shows up in the large allocation list
is the pre-allocated channel used by the logging infrastructure.

I'd like to adjust that to not require pre-allocation (or at least,
not require so much pre-allocation), but an easy memory reduction
win is to Box out the FBL portion of the record; incoming FBL
reports are incredibly rare compared to the bulk of the messages
processed by the system so it is reasonable to use an extra heap
allocation for those if it means we can save a nice chunk of
memory for the common case.