apply timerwheel batch wait to singleton wheel as well
In 4a0a4d6a1e058b1c5769f4fe32e8807372731e39 I fixed an issue where
the bucketing used in the timewheel could trick the throttling
event handler into thinking that the message had been explicitly
throttled by the event, when in reality it was just a 20ms discrepancy.

I overlooked that this same logic should also apply to the equivalent
logic for the singleton timer wheel, so that is what this commit does.