shaping.lua: explicit gc in kumo.tsa.config.monitor
The config monitor task is a long-lived task that runs
inside the same lua context for its entire lifetime.

That task obtains a reference to the current shaping data
every 30 seconds to decide whether it needs to synthesize
an update to the config epoch.

Lua's memory management cannot see the total amount of
ram consumed by the shaping data because the shaping data
is a userdata type whose ram is managed externally.
That means that lua things that this task is generally not
using many resources (~ a handful of pointers per loop iteration)
and gc doesn't kick in very aggressively.

For sites with large TSA shaping overrides, or otherwise with
very large shaping data, this can result in an accumulation
of stale shaping data in that long-lived lua context.

This commit adds an explicit gc trigger before the task
sleeps on each iteration so that we can release those references.