rfc5321/client: handle obscure DATA case in rfc2920
The pipelining spec allows for a server to proceed with a transaction
in which there were no valid recipients and validly enter the DATA
portion of the transaction.

I suspect that this might be responsible for the occasional protocol
synchronization issue that we've seen.

Handle that case per the recommendations in the RFC by sending
a dummy DATA in the case that some of the commands failed by
DATA succeeded.