Add unix domain socket support in smtp dispatcher
This commit allows setting the queue config to use a unix domain socket
path:

```
        protocol = {
          smtp = {
            mx_list = { '/some/path' },
          },
        },
```

and the smtp dispatcher will attempt to connect to it via a unix domain
socket.

It is anticipated that this will be used together with an egress path
config set to `use_lmtp = true`.

I've tested the ability to attempt to connect, but I don't have a
convenient implementation of SMTP/LMTP to test against here, so all I've
tested if that we can successfully fail to connect to a unix domain
socket.

refs: https://github.com/KumoCorp/kumomta/issues/267