mailparsing: add unit test for multi-line CD filename
A user reported that one of their users was having trouble
with a message that seemed to point at a multi-line Content-Disposition
header with some specific content.

This commit adds a unit test that shows that the issue was not due
solely to just this header, because it successfully parses.

Their issue is something else that is yet to be understood.