throttle: clean up redis lease logic
Move the limit script to a separate file so that we can
format it with stylua.

Make the debug logic conditional on a DEBUG variable so
that it has no real overhead in the common case, but we can
turn it back up when troubleshooting.  This is unfortunately
not a runtime option, so we'll need to recompile to activate
the logic, but it at least means that we can preserve it
until we need it.

Since ZRANGEBYSCORE is deprecated in redis 6.2 and up, make
our use of it vs. ZRANGE BYSCORE conditional on the redis
server version.