mailparsing: fix qp_encode with UTF-8 subsequences containing 0x20
With a specific farsi input string, there are one or more codepoints
that encode with 0x20 as part of their subsequence.  The qp_encode
function would confuse those with a space and produce the wrong
output.

This commit fixes that by having the encoder iterate by codepoint
rather than by byte, and then emit the appropriate byte sequences
from there.