smtp_server: allow 8bit envelopes
We'll accept these in MAIL FROM and RCPT TO, and leave it to the
client to indicate whether 8bit is in use.

This is partially supporting 8BITMIME and SMTPUTF8, without
explicitly claiming support for it via the EHLO capabilities.

This is probably sufficient to approximate support for sites that are
willing to try it and advertise those extensions for themselves via
https://docs.kumomta.com/reference/events/smtp_server_ehlo/

```lua
kumo.on('smtp_server_ehlo', function(domain, conn_meta, extensions)
  table.insert(extensions, 'SMTPUTF8')
  table.insert(extensions, '8BITMIME')
  return extensions
end)
```

refs: https://github.com/KumoCorp/kumomta/issues/327