policy_utils.equals now uses pairs on userdata
In an upcoming change, I want to allow comparing userdata
and tables that serialize the same way, rather than just
using the current logic which decides that a table and
a userdata can never be equal.

This commit removes that check and instead continues through
to iterate the keys/values of both operands regardless of
whether the container is a table or userdata.