config: from_lua_value will now fallback to __pairs
from_lua_value is used to attempt to deserialize a generic lua
value into a concrete Rust type that is known at compile time.

This uses some support logic inside mlua to allow:

* If it is a userdata created by mlua, mlua will try to retrieve
  that concrete rust type if the type of the userdata exactly
  matches.
* otherwise: it will use serde to deserialize the lua value
  into the rust type. mlua's deserializer cannot process
  userdata, so if we still have on at this point, the attempt
  to convert it will fail.

That's mostly fine, but our memoize crate will start returning a proxy
userdata type in a future commit, and since that proxy type will never
match the concrete Rust type that is desired by the above logic, it will
become impossible to pass memoized values without some kind of explicit
preprocessing.

What this commit does is add implicit preprocessing of the value: if we
cannot use the default mlua userdata cast, then, if the value is a
userdata, we will iterate over its fields to build a native lua
representation of that value. We also apply this translation by
traversing tables and converting their values, recursively.

While this is "preprocessing", it is only applied in case the regular
from_lua_value logic in mlua (the fast path) fails, so this is fallback
preprocessing rather than eager preprocessing, which should be
performance neutral in the common hot code paths.