memoize: improve performance of large lua tables
I was recently investigating slowness around the use of
the queue helper's `apply` method.

What was notable about this was that for this particular deployment,
the `apply` method should have been effectively a NOP but it had
a notable slowdown on the aggregate throughput.

The slowdown was because the queue.toml file was quite large, and
because our memoization cache was deep-copying the table out of the
cache on a cache hit.

What this commit does is introduce a proxy type to provide a
copy-on-write userdata value to act as a stand-in for the raw data.

What this means is that referencing the memoized value is now much
cheaper, and only the code that needs to muta the returned value (such
as in the listener domain helper) pays the cost of a deep copy, and even
then, it only pays that cost for the values that it has referenced out
of the cached value.

Prior to these changes, loading ~2000 line queue.toml file looked
like this:

initial apply: 0.006792s (which loads the data as well)
apply:         0.000576s (cached)

with these changes the number look like this:

initial apply: 0.005616s (which loads the data as well)
apply:         0.000015s (cached)

which is ~40x faster for this particular set of data.