rfc5321 client: use single buffer for pipeline send
Previously, when using pipelining, we'd still send the individual
commands one by one, resulting in potentially multiple packets over the
wire, then we'd read the command responses back.

This was really only taking advantage of half of what pipelining offers,
by avoiding the read stalls between each of the pipelined commands.

What this commit does is to pre-format those writes into a single buffer
so that we have a single logical send which stands a good chance of
being transmitted as a single packet on the wire.  (TLS may complicate
this in practice).

This should make a marginal improvement in latency, but it may also help
to address some flakiness in the integration tests that we part of
df8bb0f3aeee510fdd72db11851481f5e00404bf, where we sometimes see the
disconnect without reading the 421 response back.  My theory for those
is that, due to noisy neighbors on the CI system, we stand a good chance
of hitting a write error when writing out the later commands in the
pipeline, after the one that triggered a 421 in the peer, if that has
had time to propagate the shutdown back to us already.