rfc5321 client: break socket in more IO error cases
For the read code paths: if we experience an IO error or timeout,
ensure that we break the socket so that we don't try to reuse it
in the future, like we do for (most of) the write code paths.

For the write code paths, there were a couple of missing spots
where we didn't guarantee to break the socket, so ensure that
we do by centralizing all writes through a newly factored function
for that purpose.

Ensure that we log a diagnostic to the tracer for IO and timeout errors.

Generate a Closed event for the tracer whenever we break the socket.

Remove ClientError::Io, as we want to standardize on NotConnected
for the IO error case.  This highlighted that flushing the socket
could also generate an IO error, so add the same sort of diagnostic
trace, socket break and error code mapping there for that as well.