smtp_dispatcher: fix incorrect reuse of closed connection
In the smtp client we internally map IO and timeout errors into
a NotConnected state and break its internal socket, prevent it
from trying to talk to a broken session.

In the dispatcher layer that sits above it, when we consider
whether we need to open a connection, we look only to see if
we have a connection open or not.

If that connection is in a broken state we would keep trying to
use it, because we didn't consider that it might have such
an internal broken state.

This commit fixes up that logic so that we realize that it is
broken, and allow us to move to the next host in the connection
plan.