tidy up chrono timedelta math
I'm investigating an issue where a block of messages appear to be stuck
in a scheduled state.  In looking at the math here, I'm wondering if it
is possible for some overflow to happen when converting from chrono
to std time types, so adjust the logic to be less prone to such a thing,
if it is occurring.