smtp_dispatcher: avoid categorizing reconnects as "no more hosts"
This adds a way for the lower level dispatcher impl to communicate
back up to the higher level that the session is done and can be
closed out without treating it as a connection failure.

For SMTP the criteria for this is that, if we ever got as far
as MAIL FROM, then we consider that the connection was established
OK and that means that if we subsequently reach a point where we need to
connect again, then it must be due to an earlier policy issue that was
already recorded as a transient failure.

In that circumstance, if we have not more hosts to connect to, we
will indicate that we terminated "OK".

This works because the upper level checks for this state before
asking the session to connect again.

If there are more addresses that could be tried, we will report a
connection failure for the next message picked up by the session;
this commit only changes the behavior for the case where there
are no more addresses left in the current connection plan.