queue: fix immediately-due requeue race condition
We've been trying to run down an issue where a subset of messages
can get stuck in the scheduled queue. This has only occured in
a single deployment as far as we know; that deployment makes
use of the `requeue_message` event handler.

The identifying characteristics of the bad state are:

* `kcli inspect-sched-q` returns information about messages where
  the due time is less than the `now` field in the inspect result.
* The scheduled queue depth for that queue is flat

I've been scratching my head over this, but I *think* the issue
is a race condition during insertion.  The potential race was:

* [thread a] insert message to SINGLETON_WHEEL that is due immediately
* [wheel maintainer thread] pop due message from SINGLETON_WHEEL
* [wheel maintainer thread] verify that message is tracked by
  queue-local hash <fail>
* [thread a] add message to queue-local hash, bump scheduled queue
  counters

Those immediately due messages would sit there until a rebind forces
them to be reconsidered, or until kumod was restarted (with the same
overall effect).

The race condition could only trigger when:

* The message(s) being inserted are immediately due
* It is some multiple of 3 seconds since the singleton wheel was
  started; that is how often it awakes to process the scheduled
  queue timer wheel.

This commit re-orders things such that the queue-local hash is updated
prior to inserting into SINGLETON_WHEEL. The ordering is enforced
by the containing mutex.