rocks: add limit_concurrent_(stores|loads|removes) options
Allow enabling a semaphore to specify concurrency constraints
for each of the spool operations.

This can help to avoid spawning hundreds of tokio blocking threads
that will just sit waiting for rocksdb to complete; the main saving
here is memory, as each blocking thread has a larger amount of
stack sitting idle than the per-task memory overhead that waits
for the semaphore.