shaping.lua: ignore Rebind and Delayed log record types
Cut out rebind and delayed log records from the feed to the TSA
daemons.  These don't make sense to have TSA rules applied to
them, and in some configurations (such as deferred injection)
these records can 2x the volume of records being processed.

With pre-filtering enabled those records won't make it to the TSA
daemon, but we would still need to run those records through the
rules to realize that.

With this commit we can cut these out in O(1) and minimize the
local CPU cost of that.