queue+ready_queue: refactor: remove StdMutex type alias
This came from an iteration when we had both tokio and std mutexes in
this file.  Since then we've eliminated the former and traded out for
parking_lots mutexes.

As a result, the StdMutex label for this is quite misleading.

Remove the type alias and just use the name of the actual type.

This should have zero functional impact as it is just removing
a type alias.