logging: create sub-directory structure if needed
previously, we assumed that only the parent log_dir might be the thing
that would need to be created on demand.

If you configure per-record log files with subdirectories in them, and
the indicated directory didn't already exist, then we'd error out.

This commit will catch the error in that case and try to fill out the
directory structure on demand, which makes things a bit more convenient.