logging: skip error if log dir doesn't exist on startup
On startup, we walk the contents of a log directory to mark the
individual files as "done" (remove the w bit).

If that directory doesn't exist, because the user configured
a sub-directory hierarchy for it beyond what was explicitly
configured in the log_dir, then we would print an error
message to the journal.

That particular case is harmless: if the dir doesn't exist,
then it has no files to mark as done, so there is nothing to do.

Adjust to avoid printing an error in that case.