shaping: micro-optimize get_egress_path_config impl w/ providers
We make two passes over the provider data; the first to get
the base provider stuff which can theoretically match multiple
provider records, then a second to apply provider-source data
over the top of that.

That second pass should ideally hit the DNS cache for any of
the providers that we checked in the first pass, but it is
theoretically possible for entries to be pushed out of the
cache or expire in the intervening time.

Let's just keep a local note of what we want to include to avoid
that possibility; it also shaves off some micro cpu overhead to
decide what should apply.