queue: make QueueManager::resolve cancel safe
Previously, this code implemented work deduplication through the
use of a simple Notify object to signal waiters after the primary
task had completed.

I'd like to apply a timeout to the resolve function but the simplistic
signalling mechanism will be prone to leaving pending waiters stalled
if the task that is responsible for filling out the slot is simply
never polled again once the timeout is reached--it doesn't behave
well in the cancellation case.

This commit switches to a more robust signalling mechanism that
uses a Semaphore to ensure that all waiters have a chance to wake
up and respond to a cancelled task on the same queue name.

While in here, I put in a negative cache so that in a persistent
error scenario, we're not just subjecting the queue of waiters
to re-attempt something that is likely to fail in a similar
time frame.